+++ src/main/web/report/finalist/finalist.js
@@ -28,19 +28,22 @@ const finalist_module = {}
     let _teams;
     let _categories;
     let _tournament;
+    let _runningHead2Head;
     let _divisions;
     let _currentDivision;
     let _numTeamsAutoSelected;
     let _scheduleParameters;
     let _categoriesVisited;
-    let _currentCategoryName; // category to display with numeric.html
-    let _playoffSchedules;
+    let _currentCategoryName; // category to display with numeric.html    
+    let _playoffSchedules; // group name -> PlayoffSchedule 
+    let _finalistGroups; // group name -> FinalistGroup
     let _schedules; // awardGroup -> list of FinalistDBRow
 
     function _init_variables() {
         _teams = {};
         _categories = {};
         _tournament = null;
+        _runningHead2Head = true;
         _divisions = [];
         _currentDivision = null;
         _numTeamsAutoSelected = 1;
@@ -48,6 +51,7 @@ const finalist_module = {}
         _categoriesVisited = {};
         _currentCategoryName = null;
         _playoffSchedules = {};
+        _finalistGroups = {};
         _schedules = {};
     }
 
@@ -58,6 +62,7 @@ const finalist_module = {}
         fllStorage.set(STORAGE_PREFIX, "_teams", _teams);
         fllStorage.set(STORAGE_PREFIX, "_categories", _categories);
         fllStorage.set(STORAGE_PREFIX, "_tournament", _tournament);
+        fllStorage.set(STORAGE_PREFIX, "_runningHead2Head", _runningHead2Head);
         fllStorage.set(STORAGE_PREFIX, "_divisions", _divisions);
         fllStorage.set(STORAGE_PREFIX, "_currentDivision", _currentDivision);
         fllStorage.set(STORAGE_PREFIX, "_numTeamsAutoSelected",
@@ -66,6 +71,7 @@ const finalist_module = {}
         fllStorage.set(STORAGE_PREFIX, "_categoriesVisited", _categoriesVisited);
         fllStorage.set(STORAGE_PREFIX, "_currentCategoryName", _currentCategoryName);
         fllStorage.set(STORAGE_PREFIX, "_playoffSchedules", _playoffSchedules);
+        fllStorage.set(STORAGE_PREFIX, "_finalistGroups", _finalistGroups);
 
         fllStorage.set(STORAGE_PREFIX, "_schedules", _schedules);
     }
@@ -87,6 +93,10 @@ const finalist_module = {}
         value = fllStorage.get(STORAGE_PREFIX, "_tournament");
         if (null != value) {
             _tournament = value;
+        }
+        value = fllStorage.get(STORAGE_PREFIX, "_runningHead2Head");
+        if (null != value) {
+            _runningHead2Head = value;
         }
         value = fllStorage.get(STORAGE_PREFIX, "_divisions");
         if (null != value) {
@@ -115,6 +125,10 @@ const finalist_module = {}
         value = fllStorage.get(STORAGE_PREFIX, "_playoffSchedules");
         if (null != value) {
             _playoffSchedules = value;
+        }
+        value = fllStorage.get(STORAGE_PREFIX, "_finalistGroups");
+        if (null != value) {
+            _finalistGroups = value;
         }
 
         value = fllStorage.get(STORAGE_PREFIX, "_schedules");
@@ -271,6 +285,18 @@ const finalist_module = {}
     }
 
     /**
+     * Make sure that the start and end times are JSJoda.LocalTime objects.
+     */
+    function _fixFinalistGroup(group) {
+        if (null != group.startTime && !(group.startTime instanceof JSJoda.LocalTime)) {
+            group.startTime = JSJoda.LocalTime.parse(group.startTime);
+        }
+        if (null != group.endTime && !(group.endTime instanceof JSJoda.LocalTime)) {
+            group.endTime = JSJoda.LocalTime.parse(group.endTime);
+        }
+    }
+
+    /**
      * Make sure that the start is a JSJoda.LocalTime object.
      */
     function _fixScheduleParameters(params) {
@@ -369,6 +395,14 @@ const finalist_module = {}
 
     finalist_module.setTournament = function(tournament) {
         _tournament = tournament;
+    };
+
+    finalist_module.setRunningHead2Head = function(value) {
+        _runningHead2Head = value;
+    };
+
+    finalist_module.getRunningHead2Head = function() {
+        return _runningHead2Head;
     };
 
     /**
@@ -471,6 +505,51 @@ const finalist_module = {}
         existing.endTime = time;
         _playoffSchedules[division] = existing;
     };
+
+    /**
+     * The finalist groups. Note that the times may be null. 
+     *
+     * @return key= name, value=FinalistGroup
+     */
+    finalist_module.getFinalistGroups = function() {
+        return _finalistGroups;
+    };
+
+    /**
+     * @param newGroups the new finalist groups
+     */
+    finalist_module.setFinalistGroups = function(newGroups) {
+        _finalistGroups = newGroups;
+    }
+
+    /**
+     * Find the finalist group for the specified time.
+     * @return FinalistGroup or null when the team is not a member of any group
+     */
+    finalist_module.getFinalistGroupForTeam = function(team) {
+        for (const [_, group] of Object.entries(finalist_module.getFinalistGroups())) {
+            for (const judgingGroup of group.judgingGroups) {
+                if (team.judgingGroup == judgingGroup) {
+                    return group;
+                }
+            }
+        }
+        return null;
+    };
+
+    /**
+     * Get all judging groups from all teams.
+     * @return a possibly empty set of judging groups
+     */
+    finalist_module.getJudgingGroups = function() {
+        const judgingGroups = new Set();
+        for (const [_, team] of Object.entries(_teams)) {
+            if (team.judgingGroup) {
+                judgingGroups.add(team.judgingGroup);
+            }
+        }
+        return judgingGroups;
+    }
 
     /**
      * Get the room for the specified category and division.
@@ -956,7 +1035,8 @@ const finalist_module = {}
                         if (!scheduled
                             && !finalist_module.isTimeslotBusy(slot, category.name)
                             && !finalist_module.isTeamInTimeslot(slot, teamNum)
-                            && !finalist_module.hasPlayoffConflict(team, slot)) {
+                            && !finalist_module.hasPlayoffConflict(team, slot)
+                            && !finalist_module.outsideFinalistGroup(team, slot)) {
                             finalist_module.addTeamToTimeslot(slot, category.name, teamNum);
                             scheduled = true;
                         }
@@ -968,7 +1048,9 @@ const finalist_module = {}
 
                         nextTime = newRow.endTime;
 
-                        if (!finalist_module.hasPlayoffConflict(team, newRow)) {
+                        const playoffConflict = finalist_module.hasPlayoffConflict(team, newRow);
+                        const outsideFinalistGroup = finalist_module.outsideFinalistGroup(team, newRow);
+                        if (!playoffConflict && !outsideFinalistGroup) {
                             scheduled = true;
                             finalist_module.addTeamToTimeslot(newRow, category.name, teamNum);
                         }
@@ -1015,10 +1097,15 @@ const finalist_module = {}
      * @param team
      *          Team object
      * @param slot
-     *          Timeslot object
-     * @returns true or false
+     *          Timeslot object or FinalistDBRow object (needs time and endTime properties)
+     * @returns true if there is a conflict
      */
     finalist_module.hasPlayoffConflict = function(team, slot) {
+        if (!finalist_module.getRunningHead2Head()) {
+            // cannot have a conflict if not running head to head
+            return false;
+        }
+
         let conflict = false;
         for (const bracketName of team.playoffDivisions) {
             const playoffSchedule = _playoffSchedules[bracketName];
@@ -1040,6 +1127,11 @@ const finalist_module = {}
      * @returns true or false
      */
     finalist_module.slotHasPlayoffConflict = function(playoffSchedule, slot) {
+        if (!finalist_module.getRunningHead2Head()) {
+            // cannot have a conflict if running head to head
+            return false;
+        }
+
         const start = playoffSchedule.startTime;
         const end = playoffSchedule.endTime;
         if (null != start && null != end) {
@@ -1051,11 +1143,61 @@ const finalist_module = {}
         return false;
     };
 
+
     /**
-     * The schedule parameters.
-     *
-     * @param currentDivision the award group to get the schedule start time for
+     * Check if the specified time slot is outside the finalist group schedule for the team based on their judging group.
+     * 
+     * @param team Team object
+     * @param slot Timeslot object
+     * @returns true if outside the schedule for their group
      */
+    finalist_module.outsideFinalistGroup = function(team, slot) {
+        if (finalist_module.getRunningHead2Head()) {
+            // finalist groups are only used when not running head to head
+            return false;
+        }
+
+        // allow times after the last group
+        let latestGroupEnd = null;
+        for (const [_, group] of Object.entries(finalist_module.getFinalistGroups())) {
+            if (group.endTime) {
+                if (!latestGroupEnd) {
+                    latestGroupEnd = group.endTime;
+                } else if (latestGroupEnd.isBefore(group.endTime)) {
+                    latestGroupEnd = group.endTime;
+                }
+            }
+        }
+        if (!latestGroupEnd) {
+            // no latest, must not be groups, so just allow it
+            return false;
+        }
+        if (!slot.time.isBefore(latestGroupEnd)) {
+            // slot is after last group
+            return false;
+        }
+
+        const group = finalist_module.getFinalistGroupForTeam(team);
+        if (null == group) {
+            // no group, let's assume it's ok
+            return false;
+        }
+        if (null != group.startTime && null != group.endTime) {
+            if (slot.time.isBefore(group.startTime)) {
+                return true;
+            }
+            if (group.endTime.isBefore(slot.endTime)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+ * The schedule parameters.
+ *
+ * @param currentDivision the award group to get the schedule start time for
+ */
     finalist_module.getScheduleParameters = function(currentDivision) {
         let prevValue = _scheduleParameters[currentDivision];
         if (undefined == prevValue) {
@@ -1323,6 +1465,47 @@ const finalist_module = {}
     };
 
     /**
+     * Load the finalist groups from the server.
+     * 
+     * @return promise to execute
+     */
+    finalist_module.loadFinalistGroups = function() {
+        return fetch("../../api/FinalistGroups").then(checkJsonResponse).then(function(data) {
+            for (const [name, group] of Object.entries(data)) {
+                _fixFinalistGroup(group);
+                _finalistGroups[name] = group;
+            }
+        });
+    };
+
+    /**
+     * Upload the finalist groups to the server.
+     * 
+     * @param successCallback called with the server result on success
+     * @param failCallback called with the server result on failure
+     * @return promise to execute
+     */
+    finalist_module.uploadFinalistGroups = function(successCallback, failCallback) {
+        const groupsToUpload = {};
+        for (const [name, group] of Object.entries(_finalistGroups)) {
+            if (null != group.startTime && null != group.endTime) {
+                groupsToUpload[name] = group;
+            }
+        }
+
+        return uploadJsonData("../../api/FinalistGroups", "POST", groupsToUpload)
+            .then(checkJsonResponse).then(function(result) {
+                if (result.success) {
+                    successCallback(result);
+                } else {
+                    failCallback(result);
+                }
+            }).catch(function(result) {
+                failCallback(result);
+            });
+    };
+
+    /**
       * Load the finalist schedule parameters from the server.
       * 
       * @return promise to execute
@@ -1462,6 +1645,17 @@ const finalist_module = {}
             for (const bracket of data) {
                 finalist_module.addPlayoffDivision(bracket);
             }
+        });
+    };
+
+    /**
+      * Load the tournament parameters from the server.
+      * 
+      * @return promise to execute
+      */
+    finalist_module.loadTournamentParameters = function() {
+        return fetch("../../api/TournamentParameters").then(checkJsonResponse).then(function(data) {
+            finalist_module.setRunningHead2Head(data.runningHead2Head);
         });
     };
 
@@ -1679,6 +1873,12 @@ const finalist_module = {}
         waitList1.push(playoffBracketsPromise);
 
 
+        const tournamentParametersPromise = finalist_module.loadTournamentParameters();
+        tournamentParametersPromise.catch(function() {
+            failCallback("Tournament Parameters");
+        })
+        waitList1.push(tournamentParametersPromise);
+
         Promise.all(waitList1).then(function(_) {
             // everything else can be loaded in parallel
             const waitList = [];
@@ -1722,6 +1922,12 @@ const finalist_module = {}
             failCallback("Playoff Schedules");
         })
         waitList.push(playoffSchedulesPromise);
+
+        const finalistGroupsPromise = finalist_module.loadFinalistGroups();
+        finalistGroupsPromise.catch(function() {
+            failCallback("Finalist groups");
+        })
+        waitList.push(finalistGroupsPromise);
 
         const finalistParamsPromise = finalist_module.loadFinalistScheduleParameters();
         finalistParamsPromise.catch(function() {